feat: store for global source of truth
React is an UI library based on reactive behavior. Due its
ephemeral approach of components, there are some options for
storing state in a more global approach.

Between the available options, the chosen was Redux.

Some stores were impelemented for dealing with states and changes:
- Auth: stores Authentication information
- User: stores User related information

Its important to note that for Sensitive information such as the
`accessToken`, it was stored inside the Secure Storage and not at
Redux state nor Async Storage.